fix: convert remaining float timestamps to int milliseconds
Align with Rust wire format — all timestamps are int milliseconds
since epoch, not float seconds. Fixes: HalfBlock.timestamp,
InteractionRecord.timestamp, TransportMessage.timestamp,
SQLiteBlockStore revoked_at/created_at, envelope encoding/decoding,
trust decay functions, and related tests.

+++ src/trustchain/block.py
@@ -25,7 +25,7 @@ class HalfBlock:
     block_hash: str
     interaction_type: str
     outcome: str
-    timestamp: float
+    timestamp: int
 
 
 class Block:
